Added a function checkBoardRes to retrieve the dimensions of the terminal.
The purpose is to get the terminal window's dimensions and then make them available as variables for width and height. Process.stdout.columns and process.stdout.rows are being used to get the number of columns and rows in the terminal. The return object "return { width, height };" will make sure to return an object consisting of two properties. const { width, height } = checkBoardRes(); calls the function and uses a destructuring assignment to extract width and height directly from the object returned. So, I am retrieving the height and width of the terminal window, then making them available as constants.