Added minimum resolution check to ensure that the game does not start if the current console window is below a set minimum resolution.
Created an if statement to check whether the terminal is large enough to run the game. It checks whether the terminal window is at least 110 columns wide and 28 rows heigh in order to render successfully. So, if height is less than 28 and width is less than 110, the game will not run and an error is thrown.