Added select language option to menu.
Added a new menu item called "Select Your Preferred Language". Then inside the action function in this new menu item, I called a function called selectLanguage that takes a callback function as an argument. Then based on what the player chooses to do, the callback function sets a language variable which is global and makes sure to log the message to confirm what language the player chooses.