Added language selection functionality to menu.
Created and implemented a function to allow the players to choose between two languages, English and Norwegian (Norsk). ClearScreen() makes sure the screen is cleared before showing the options and the readline interface makes sure that the user's choice is being captured. When choosing either "1" or "2"  the callback function will be executed to update the global variable.