Added functionality to change language based on the player's input.
Added rl.question to ask the player to type 1 or 2 in the terminal and based on what number the player chooses to type, the global variable called "language" will update. Also added if statement so if player press 1, language is set to English and if 2, the language is set to Norwegian and if not 1 or 2, there will be an error message. Also tried to make a translate function that makes sure the selected language is displayed. Then the callback function that is passed to selectLanguage is being called on with the choosen language.